+++ src/functions.php
@@ -6,6 +6,16 @@
  */
 
 /**
+ * Gets page instance.
+ *
+ * @param  string $page_slug Page slug.
+ * @return \PuppyFW\Page
+ */
+function puppyfw_get_page( $page_slug ) {
+	$pages = puppyfw()->get_pages();
+}
+
+/**
  * Filters save value of checkbox field.
  *
  * @param  mixed $value Checkbox save value.